Avance en la capa bl con login y registrar estudiante

+++ internship/src/main/java/edu/bo/ucb/sis213/internship/dao/ApproveCompanyRepository.java
@@ -2,6 +2,6 @@ package edu.bo.ucb.sis213.internship.dao;
 
 import edu.bo.ucb.sis213.internship.entity.ApproveCompany;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface ApproveCompanyRepository extends JpaRepository<ApproveCompany,Long>{
+public interface ApproveCompanyRepository extends JpaRepository<ApproveCompany,Integer>{
 
 }
