+++ internship/src/main/java/edu/bo/ucb/sis213/internship/dao/ApproveInternshipRepository.java
@@ -2,5 +2,5 @@ package edu.bo.ucb.sis213.internship.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import edu.bo.ucb.sis213.internship.entity.ApproveInternship;
-public interface ApproveInternshipRepository extends JpaRepository<ApproveInternship,Long>{
+public interface ApproveInternshipRepository extends JpaRepository<ApproveInternship,Integer>{
 }
