+++ internship/src/main/java/edu/bo/ucb/sis213/internship/dao/CampusRepository.java
@@ -2,5 +2,5 @@ package edu.bo.ucb.sis213.internship.dao;
 
 import edu.bo.ucb.sis213.internship.entity.Campus;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface CampusRepository extends JpaRepository<Campus,Long>{
+public interface CampusRepository extends JpaRepository<Campus,Integer>{
 }
