+++ internship/src/main/java/edu/bo/ucb/sis213/internship/dao/CareerRepository.java
@@ -3,5 +3,5 @@ package edu.bo.ucb.sis213.internship.dao;
 import edu.bo.ucb.sis213.internship.entity.Career;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CareerRepository extends JpaRepository<Career,Long> {
+public interface CareerRepository extends JpaRepository<Career,Integer> {
 }
