+++ internship/src/main/java/edu/bo/ucb/sis213/internship/dao/CarrerInternshipRepository.java
@@ -2,5 +2,5 @@ package edu.bo.ucb.sis213.internship.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import edu.bo.ucb.sis213.internship.entity.CarrerInternship;
-public interface CarrerInternshipRepository extends JpaRepository<CarrerInternship,Long>{
+public interface CarrerInternshipRepository extends JpaRepository<CarrerInternship,Integer>{
 }
