+++ internship/src/main/java/edu/bo/ucb/sis213/internship/dao/DocumentRepository.java
@@ -2,5 +2,5 @@ package edu.bo.ucb.sis213.internship.dao;
 import edu.bo.ucb.sis213.internship.entity.Document;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocumentRepository extends JpaRepository<Document,Long>{
+public interface DocumentRepository extends JpaRepository<Document,Integer>{
 }
