+++ internship/src/main/java/edu/bo/ucb/sis213/internship/dao/FacultyRepository.java
@@ -2,5 +2,5 @@ package edu.bo.ucb.sis213.internship.dao;
 import edu.bo.ucb.sis213.internship.entity.Faculty;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacultyRepository extends JpaRepository<Faculty,Long>{
+public interface FacultyRepository extends JpaRepository<Faculty,Integer>{
 }
