+++ internship/src/main/java/edu/bo/ucb/sis213/internship/dao/NotificationContactRepository.java
@@ -1,5 +1,5 @@
 package edu.bo.ucb.sis213.internship.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import edu.bo.ucb.sis213.internship.entity.NotificationContact;
-public interface NotificationContactRepository extends JpaRepository<NotificationContact,Long>{
+public interface NotificationContactRepository extends JpaRepository<NotificationContact,Integer>{
 }
