+++ internship/src/main/java/edu/bo/ucb/sis213/internship/dao/NotificationStudentRepository.java
@@ -1,5 +1,5 @@
 package edu.bo.ucb.sis213.internship.dao;
 import edu.bo.ucb.sis213.internship.entity.NotificationStudent;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface NotificationStudentRepository extends JpaRepository<NotificationStudent,Long>{
+public interface NotificationStudentRepository extends JpaRepository<NotificationStudent,Integer>{
 }
