+++ internship/src/main/java/edu/bo/ucb/sis213/internship/dao/SectorCompanyRepository.java
@@ -1,5 +1,5 @@
 package edu.bo.ucb.sis213.internship.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import edu.bo.ucb.sis213.internship.entity.SectorCompany;
-public interface SectorCompanyRepository extends JpaRepository<SectorCompany,Long>{
+public interface SectorCompanyRepository extends JpaRepository<SectorCompany,Integer>{
 }
