+++ internship/src/main/java/edu/bo/ucb/sis213/internship/dao/SectorRepository.java
@@ -1,5 +1,5 @@
 package edu.bo.ucb.sis213.internship.dao;
 import edu.bo.ucb.sis213.internship.entity.Sector;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface SectorRepository extends JpaRepository<Sector,Long>{
+public interface SectorRepository extends JpaRepository<Sector,Integer>{
 }
